+++ resources/components/laraknife/icons/exchange-record.blade.php
@@ -1,0 +1,2 @@
+@props(['module' => '', 'method' => 'filechange', 'no' => ''])
+<a href="/{{$module}}-{{$method}}/{{$no}}"><x-laraknife.icons.exchange class="text-primary" /></a>