+++ resources/components/laraknife/icons/exchange.blade.php
@@ -1,0 +1,2 @@
+@props(['class' => ''])
+<i class="bi bi-arrow-up-circle {{$class}}"></i>