+++ resources/helpers/Builder.php
@@ -3,29 +3,12 @@ namespace App\Helpers;
 function logToFile(string $message){
     file_put_contents('/tmp/builder.log', $message . "\n", FILE_APPEND);
 }
-$dir = __DIR__;
-logToFile("= Builder.import: $dir");
-if ( ($ix = strpos($dir, 'vendor/')) !== false){
-    $dir = substr($dir, 0, $ix + 7);
-    logToFile("changed: $dir");
-}
-while( ($ix = strrpos($dir, '/')) !== false){
-    $dir = substr($dir, 0, $ix);
-    if (is_dir("$dir/vendor")){
-        logToFile("found-1: $dir/vendor");
-        require "$dir/vendor/autoload.php";
-        require_once "StringHelper.php";
-        require_once "OsHelper.php";
-        $dir = null;
-        break;
-    }
-}
-if ($dir != null && is_dir('/home/ws/php/gadeku/vendor')){
-    $dir = '/home/ws/php/gadeku';
-    logToFile("found-2: $dir/vendor");
-    require "$dir/vendor/autoload.php";
+$_dir = getcwd() . '/vendor';
+if (is_dir($_dir)){
+    require "$_dir/autoload.php";
     require_once "StringHelper.php";
     require_once "OsHelper.php";
+    $_dir = null;
 }
 use App\Helpers\OsHelper;
 use App\Helpers\StringHelper;