+++ resources/helpers/FileHelper.php
@@ -1,5 +1,6 @@
 <?php
 namespace App\Helpers;
+use Illuminate\Http\Request;
 
 class FileHelper{
     /**
@@ -14,8 +15,9 @@ class FileHelper{
     }
    /**
      * Builds the relative path for the upload file storage.
+     * @param string|\DateTime $date the creation date of the file: defines the path name
      */
-    public static function buildFileStoragePath($date=null): string{
+    public static function buildFileStoragePath($date): string{
         $date ??= new \DateTime();
         if ($date instanceof \DateTime){
             $rc = $date->format('Y') . '/' . $date->format('m');
@@ -28,8 +30,9 @@ class FileHelper{
     /**
      * Deletes the uploaded file.
      * @param string $filename the name to rename (without path)
+     * @param string|\DateTime $date the creation date of the file: defines the path name
      */
-    public static function deleteUploadedFile(string $filename, \DateTime $date){
+    public static function deleteUploadedFile(string $filename, $date){
         $name = storage_path() . '/app/public/' . FileHelper::buildFileStoragePath($date) . '/' . $filename;
         \unlink($name);
     }
@@ -47,11 +50,29 @@ class FileHelper{
      * That is needed because the filename contains the primary key. This is known after storing the record.
      * @param string $oldName the name to rename (without path)
      * @param string $newName the target name
+     * @param string|\DateTime $date the creation date of the file: defines the path name
      */
-    public static function renameUploadedFile(string $oldName, string $newName){
-        $storage = storage_path() . '/app/public/' . FileHelper::buildFileStoragePath() . '/';
+    public static function renameUploadedFile(string $oldName, string $newName, $date){
+        $storage = storage_path() . '/app/public/' . FileHelper::buildFileStoragePath($date) . '/';
         $old = $storage  . $oldName;
         $new = $storage . $newName;
         \rename($old, $new);
+    }
+    /**
+     * Replaces the uploaded file.
+     * @param Request $request
+     * @param string $fieldname the name of the field with type "file"
+     * @param string $filename the name of the file to replace
+     * @param string|\DateTime $date the creation date of the file: defines the path name
+     */
+    public static function replaceUploadedFile(Request $request, string $fieldname, string $filename, $date){
+        $relativePath = FileHelper::buildFileStoragePath($date);
+        $request->file($fieldname)->storeAs($relativePath, $filename, 'public');
+    }
+
+    public static function storeFile(Request $request, string $fieldname, string $filename): string{
+        $relativePath = FileHelper::buildFileStoragePath(null);
+        $filePath = $request->file($fieldname)->storeAs($relativePath, $filename, 'public');
+        return $filePath;
     }
 }